Redefine array data structures

+++ src/branch_sequences.c
@@ -525,8 +525,7 @@ void get_likelihood_for_windows(char * child_sequence, int length_of_sequence, i
   }
 }
 
-//int extend_upper_part_of_window(int starting_coord, int initial_max_coord, int genome_size, int8_t * gaps_in_original_genome_space)
-int extend_upper_part_of_window(int starting_coord, int initial_max_coord, int genome_size, int * gaps_in_original_genome_space)
+int extend_upper_part_of_window(int starting_coord, int initial_max_coord, int genome_size, int8_t * gaps_in_original_genome_space)
 {
 		int max_snp_sliding_window_counter = initial_max_coord;
 		int upper_offset = 0;
@@ -543,8 +542,7 @@ int extend_upper_part_of_window(int starting_coord, int initial_max_coord, int g
 		return max_snp_sliding_window_counter;
 }
 
-//int extend_lower_part_of_window(int starting_coord, int initial_min_coord, int genome_size, int8_t * gaps_in_original_genome_space)
-int extend_lower_part_of_window(int starting_coord, int initial_min_coord, int genome_size, int * gaps_in_original_genome_space)
+int extend_lower_part_of_window(int starting_coord, int initial_min_coord, int genome_size, int8_t * gaps_in_original_genome_space)
 {
 		int lower_offset = 0;
 		int snp_sliding_window_counter = initial_min_coord;
@@ -564,16 +562,12 @@ int extend_lower_part_of_window(int starting_coord, int initial_min_coord, int g
 int get_blocks(int ** block_coordinates, int genome_size,int * snp_site_coords,int number_of_branch_snps, int window_size, int cutoff, char * original_sequence, int * snp_locations, int number_of_snps)
 {
 	// Set up the window counter with 1 value per base in the branch
-//    int8_t * window_count;
-//	window_count = (int8_t *) calloc((genome_size+1),sizeof(int8_t));
     int * window_count;
     window_count = (int *) calloc((genome_size+1),sizeof(int));
 	
 	// Integer array with location of gaps
-//    int8_t * gaps_in_original_genome_space;
-//	gaps_in_original_genome_space = (int8_t *) calloc((genome_size+1),sizeof(int8_t));
-    int * gaps_in_original_genome_space;
-    gaps_in_original_genome_space = (int *) calloc((genome_size+1),sizeof(int));
+    int8_t * gaps_in_original_genome_space;
+	gaps_in_original_genome_space = (int8_t *) calloc((genome_size+1),sizeof(int8_t));
 	int x =0;
 	for(x=0; x< number_of_snps; x++)
 	{
